+++ pytsite/semver/__init__.py
@@ -6,7 +6,7 @@ http://semver.org/
 # Public API
 from . import _error as error
 from ._api import parse_version_str, compare, last, check_conditions, to_int, parse_requirement_str, \
-    parse_condition_str, minimum, maximum
+    parse_condition_str, minimum, maximum, increment, decrement
 from ._version import Version
 
 __author__ = 'Alexander Shepetko'